fruitlist component logic to fruitlistcontainer component
